fix: DS-4, SR-2, MI-4, IS-1, DS-3 generator correctness and speedup issues
- DS-4: add include guard to struct typedef so test.c compiles as separate TU
- SR-2: fix float literal syntax in helper.c, damp polynomial penalty to prevent overflow
- MI-4: fix UNARY_MATH_FNS tuple unpacking, widen float reduce tolerance
- IS-1: restrict to quadratic/cubic layouts, raise minimum sparsity to 90%+
- DS-3: both functions noinline in helper.c, larger structs (512-1024B), access only 3 fields

All 580 variants now pass correctness tests (previously 530/555).

+++ dataset/IS_1/IS-1_v004/fast.c
@@ -1,11 +1,11 @@
-void fast_is1_v004(double *C, double *A, double *B, int m, int k, int n) {
+void fast_is1_v004(float *C, float *A, float *B, int m, int k, int n) {
     for (int i = 0; i < m; i++)
-        for (int j = 0; j < n; j++) C[i * n + j] = 0.0;
+        for (int j = 0; j < n; j++) C[i * n + j] = 0.0f;
     for (int i = 0; i < m; i++) {
         for (int p = 0; p < k; p++) {
-            if (A[i * k + p] == 0.0) continue;
+            if (A[i * k + p] == 0.0f) continue;
             for (int j = 0; j < n; j++) {
-                if (B[p * n + j] == 0.0) continue;
+                if (B[p * n + j] == 0.0f) continue;
                 C[i * n + j] += A[i * k + p] * B[p * n + j];
             }
         }
