+++ dataset/IS_1/IS-1_v009/slow.c
@@ -1,6 +1,6 @@
-void slow_is1_v009(double *y, double *A, double *x, int m, int n) {
+void slow_is1_v009(float *y, float *A, float *x, int m, int n) {
     for (int i = 0; i < m; i++) {
-        y[i] = 0.0;
+        y[i] = 0.0f;
         for (int j = 0; j < n; j++) {
             y[i] += A[i * n + j] * x[j];
         }
