+++ dataset/MI_4/MI-4_v014/fast.c
@@ -1,5 +1,5 @@
-double fast_mi4_v014(double *matrix, int rows, int cols) {
-    double total = 0;
+int fast_mi4_v014(int *matrix, int rows, int cols) {
+    int total = 0;
     for (int i = 0; i < rows; i++) {
         for (int j = 0; j < cols; j++) {
             total += matrix[i * cols + j];
